requirements.txt
appdirs==1.4.3
attrs==19.3.0
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.3
pathspec==0.6.0
Pillow==6.2.1
regex==2019.11.1
six==1.12.0
toml==0.10.0
toolz==0.10.0
torch==1.3.0.post2
torchvision==0.4.1.post2
typed-ast==1.4.0
Werkzeug==0.16.0
